requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
fsspec==2024.3.1
fusepy==3.0.1
gitdb==4.0.11
GitPython==3.1.41
idna==3.7
requests==2.31.0
setuptools==69.0.3
smmap==5.0.1
urllib3==2.2.1
wheel==0.43.0
python-dotenv==1.0.1